switch data input to planning page
