+++ src/pages/planning/index.js
@@ -2,27 +2,41 @@ import React from 'react';
 import { State } from '~src/state';
 import IcicleChart from './icicleChart';
 
+import Transactions from './transactions';
+import Accounts from './accounts';
+
 const Planning = (props) => {
   return (
     <State.Consumer>
       {(model) => (
-        <React.Fragment>
-          {/* {console.log(model.state)} */}
-          <div>
+        <div className="container mx-auto my-2 px-4 sm:px-6 lg:px-8">
+          <h1 className="text-3xl font-semibold py-3">Cash Flow</h1>
+          <div className="bg-white shadow rounded-lg divide-y divide-gray-200">
             <IcicleChart data={model.state.transactionsComputed} />
-            {model.state.transactionsComputed.map((transaction) => (
-              <div>
-                <span>{transaction.raccount}</span>
-                <span>{transaction.category}</span>
-                <span>{transaction.dailyRate.toFixed(2)}</span>
-              </div>
-            ))}
-            <pre>{JSON.stringify(model.state, null, 2)}</pre>
           </div>
-        </React.Fragment>
+          <Divider text="Transactions" />
+          <Transactions />
+          <Divider text="Accounts" />
+          <Accounts />
+        </div>
       )}
     </State.Consumer>
   );
 };
 
 export default Planning;
+
+const Divider = ({ text }) => {
+  return (
+    <div className="relative my-8">
+      <div className="absolute inset-0 flex items-center" aria-hidden="true">
+        <div className="w-full border-t border-gray-300" />
+      </div>
+      <div className="relative flex justify-center">
+        <span className="px-3 bg-gray-50 text-lg font-medium text-gray-900">
+          {text}
+        </span>
+      </div>
+    </div>
+  );
+};
